refactor(desktop): move app_test.go's tests beside the files #405 created
When #405 split app.go into transferstate.go, transfer.go and endpoints.go,
app_test.go stayed whole, so a reader looking for the cancel and close-guard
tests had to know they were filed under the app. The nine generation-tag
tests and their fakeCloser now live in transferstate_test.go; TestShareLink
joins the endpoints tests in endpoints_test.go; TestWriteTextTemp and
TestRequireRelay, the two transfer.go helpers a test can reach without a
peer, live in transfer_test.go. app_test.go keeps what still exercises
app.go: TestFilterFileArgs, the three notifyTransferFailed toast tests and
the pending-files staging tests. updatecheck_test.go is untouched. The new
files carry no GOOS token, and their notes sit below the package clause.

Pure move: the sorted go test -list output is identical (86 funcs) on
Windows and GOOS=linux, the -v PASS list is identical (85 PASS, 1 SKIP), go
doc -all -u is unchanged, frontend/wailsjs/go/main/App.d.ts is byte-identical,
and git diff --color-moved shows only package, import and header lines as new.
