+++ desktop/app_test.go
@@ -1,73 +1,10 @@
 package main
 
 import (
-	"errors"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 )
-
-// TestShareLink guards the property that matters: the room id is the transfer's
-// only secret, so it must live in the URL fragment, which browsers never send to
-// a server. Anything before the '#' does reach the server, so the room id
-// appearing there would leak it into access logs, Referer headers, and analytics.
-func TestShareLink(t *testing.T) {
-	const (
-		base   = "https://floe.one"
-		nonce  = "deadbeef"
-		roomID = "6f207790-92a6-4662-bb68-4c4059f75139"
-	)
-
-	got := shareLink(base, nonce, roomID)
-
-	if want := base + "/?s=" + nonce + "#room=" + roomID; got != want {
-		t.Errorf("shareLink = %q, want %q", got, want)
-	}
-
-	// The invariant, asserted independently of the exact shape above.
-	beforeFragment, _, found := strings.Cut(got, "#")
-	if !found {
-		t.Fatalf("shareLink = %q, want a '#' fragment", got)
-	}
-	if strings.Contains(beforeFragment, roomID) {
-		t.Errorf("room id leaked into the server-visible part %q", beforeFragment)
-	}
-	if strings.Contains(got, "?room=") {
-		t.Errorf("shareLink = %q, must not use the leaky ?room= form", got)
-	}
-	if strings.Contains(got, base+"//") {
-		t.Errorf("shareLink = %q, has a double slash after the host", got)
-	}
-}
-
-// TestWriteTextTemp verifies the text-send staging: exact content round-trip,
-// the fixed message.txt name the receiver will see, and cleanup removing the
-// temp directory.
-func TestWriteTextTemp(t *testing.T) {
-	const text = "hello floe\nline two · unicode ✓"
-
-	path, cleanup, err := writeTextTemp(text)
-	if err != nil {
-		t.Fatalf("writeTextTemp: %v", err)
-	}
-
-	if got := filepath.Base(path); got != "message.txt" {
-		t.Errorf("file name = %q, want message.txt", got)
-	}
-	got, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatalf("read back: %v", err)
-	}
-	if string(got) != text {
-		t.Errorf("content = %q, want %q", got, text)
-	}
-
-	cleanup()
-	if _, err := os.Stat(filepath.Dir(path)); !os.IsNotExist(err) {
-		t.Errorf("cleanup did not remove the temp dir")
-	}
-}
 
 // TestFilterFileArgs verifies command-line arg filtering keeps only existing
 // files/dirs and drops flags, empties, and stale paths.
@@ -89,116 +26,6 @@ func TestFilterFileArgs(t *testing.T) {
 
 	if len(got) != 2 || got[0] != file || got[1] != dir {
 		t.Errorf("filterFileArgs = %v, want [%s %s]", got, file, dir)
-	}
-}
-
-// TestCancelTransferIdle guards the Start-over path, which calls CancelTransfer
-// unconditionally: with no transfer in flight (nil curSC/curConn) it must be a
-// safe no-op, never a nil-dereference panic.
-func TestCancelTransferIdle(t *testing.T) {
-	(&App{}).CancelTransfer()
-}
-
-// fakeCloser records Close calls so the generation-tag tests can prove which
-// handles a cancel actually reached.
-type fakeCloser struct{ closed int }
-
-func (f *fakeCloser) Close() { f.closed++ }
-
-// TestCancelTransferClosesLiveHandles: a cancel must close exactly the live
-// generation's registered handles.
-func TestCancelTransferClosesLiveHandles(t *testing.T) {
-	a := &App{}
-	g := a.beginTransfer()
-	sc, conn := &fakeCloser{}, &fakeCloser{}
-	if !a.setSignaling(g, sc) || !a.setConn(g, conn) {
-		t.Fatal("setters refused the live generation")
-	}
-	a.CancelTransfer()
-	if sc.closed != 1 || conn.closed != 1 {
-		t.Fatalf("cancel closed sc=%d conn=%d times, want 1/1", sc.closed, conn.closed)
-	}
-	if a.transferActive(g) {
-		t.Fatal("generation still active after cancel")
-	}
-}
-
-// TestClearTransferOldGenerationNoOp pins the race fix: a superseded
-// goroutine's deferred clear must not wipe the live transfer's handles.
-func TestClearTransferOldGenerationNoOp(t *testing.T) {
-	a := &App{}
-	g1 := a.beginTransfer()
-	sc1, conn1 := &fakeCloser{}, &fakeCloser{}
-	a.setSignaling(g1, sc1)
-	a.setConn(g1, conn1)
-
-	g2 := a.beginTransfer()
-	sc2, conn2 := &fakeCloser{}, &fakeCloser{}
-	if !a.setSignaling(g2, sc2) || !a.setConn(g2, conn2) {
-		t.Fatal("setters refused the new live generation")
-	}
-
-	// The old goroutine exits up to 30 seconds later and runs its deferred
-	// clear. It must be a no-op against the new generation's slots.
-	a.clearTransfer(g1)
-
-	a.CancelTransfer()
-	if sc2.closed != 1 || conn2.closed != 1 {
-		t.Fatalf("cancel missed the live handles after an old clear: sc2=%d conn2=%d, want 1/1", sc2.closed, conn2.closed)
-	}
-	if sc1.closed != 0 || conn1.closed != 0 {
-		t.Fatalf("cancel closed the dead generation's handles: sc1=%d conn1=%d, want 0/0", sc1.closed, conn1.closed)
-	}
-}
-
-// TestSupersededSetterRefuses: a goroutine whose generation was superseded must
-// not be able to register handles into the live transfer's slots.
-func TestSupersededSetterRefuses(t *testing.T) {
-	a := &App{}
-	g1 := a.beginTransfer()
-	_ = a.beginTransfer() // g2 supersedes g1
-
-	stale := &fakeCloser{}
-	if a.setSignaling(g1, stale) {
-		t.Fatal("superseded setSignaling accepted a handle")
-	}
-	if a.setConn(g1, stale) {
-		t.Fatal("superseded setConn accepted a handle")
-	}
-	a.CancelTransfer()
-	if stale.closed != 0 {
-		t.Fatalf("cancel closed a handle that never entered the slots: %d, want 0", stale.closed)
-	}
-}
-
-// TestCancelledSetterRefuses: after a cancel, the cancelled generation's own
-// setters must refuse, so the goroutine bails instead of parking in WebRTC
-// setup with an orphaned connection.
-func TestCancelledSetterRefuses(t *testing.T) {
-	a := &App{}
-	g := a.beginTransfer()
-	a.CancelTransfer()
-	if a.setSignaling(g, &fakeCloser{}) {
-		t.Fatal("cancelled setSignaling accepted a handle")
-	}
-	if a.setConn(g, &fakeCloser{}) {
-		t.Fatal("cancelled setConn accepted a handle")
-	}
-}
-
-// TestCancelThenBeginReArms pins the most common real flow (Cancel, then Start
-// over): beginTransfer after a cancel must fully re-arm the machinery, which
-// depends on it resetting the cancelled flag.
-func TestCancelThenBeginReArms(t *testing.T) {
-	a := &App{}
-	_ = a.beginTransfer()
-	a.CancelTransfer()
-	g2 := a.beginTransfer()
-	if !a.setSignaling(g2, &fakeCloser{}) || !a.setConn(g2, &fakeCloser{}) {
-		t.Fatal("setters refused the fresh generation after a cancel")
-	}
-	if !a.transferActive(g2) {
-		t.Fatal("fresh generation not active after a cancel")
 	}
 }
 
@@ -241,90 +68,6 @@ func TestSupersededTransferDoesNotToast(t *testing.T) {
 	a.notifyTransferFailed(g1, "Floe - send failed")
 	if calls != 0 {
 		t.Fatalf("superseded transfer toasted %d times, want 0", calls)
-	}
-}
-
-// TestCloseBlockedLifecycle pins the close guard's predicate across the
-// transfer lifecycle, including the superseded-generation hole: an old
-// goroutine's stale deferred clear must not un-guard the live transfer.
-func TestCloseBlockedLifecycle(t *testing.T) {
-	a := &App{}
-
-	if a.closeBlocked() {
-		t.Fatal("fresh app must not block close")
-	}
-
-	g1 := a.beginTransfer()
-	if !a.closeBlocked() {
-		t.Fatal("live transfer must block close")
-	}
-
-	a.clearTransfer(g1)
-	if a.closeBlocked() {
-		t.Fatal("finished transfer must not block close")
-	}
-
-	// Supersede: g2 begins, then g1's stale deferred clear fires late.
-	g1 = a.beginTransfer()
-	g2 := a.beginTransfer()
-	a.clearTransfer(g1) // stale; must be a no-op
-	if !a.closeBlocked() {
-		t.Fatal("a superseded goroutine's stale clear un-guarded the live transfer")
-	}
-	a.clearTransfer(g2)
-	if a.closeBlocked() {
-		t.Fatal("live transfer's clear must un-guard")
-	}
-
-	// A cancelled transfer must not prompt a user who already gave up.
-	a.beginTransfer()
-	a.CancelTransfer()
-	if a.closeBlocked() {
-		t.Fatal("cancelled transfer must not block close")
-	}
-}
-
-// TestConfirmCloseCancelsLatchesAndQuits: Close anyway aborts the transfer,
-// quits exactly once, and latches allowClose one-way, so not even a LATER
-// transfer can re-block the close that is already in motion.
-func TestConfirmCloseCancelsLatchesAndQuits(t *testing.T) {
-	quits := 0
-	a := &App{quitFn: func() { quits++ }}
-
-	g := a.beginTransfer()
-	sc, conn := &fakeCloser{}, &fakeCloser{}
-	if !a.setSignaling(g, sc) || !a.setConn(g, conn) {
-		t.Fatal("setters refused for the live generation")
-	}
-
-	a.ConfirmClose()
-
-	if sc.closed != 1 || conn.closed != 1 {
-		t.Fatalf("handles closed sc=%d conn=%d, want 1 and 1", sc.closed, conn.closed)
-	}
-	if quits != 1 {
-		t.Fatalf("quit called %d times, want 1", quits)
-	}
-	if a.closeBlocked() {
-		t.Fatal("close still blocked after ConfirmClose")
-	}
-
-	// The latch is one-way by design: the app is exiting, and a transfer that
-	// somehow begins during teardown must not resurrect the guard.
-	a.beginTransfer()
-	if a.closeBlocked() {
-		t.Fatal("allowClose latch did not survive a later beginTransfer")
-	}
-}
-
-// TestConfirmCloseIdle covers "transfer finished while the dialog was open,
-// user clicks Close anyway": nothing to cancel, quit still happens.
-func TestConfirmCloseIdle(t *testing.T) {
-	quits := 0
-	a := &App{quitFn: func() { quits++ }}
-	a.ConfirmClose()
-	if quits != 1 {
-		t.Fatalf("quit called %d times, want 1", quits)
 	}
 }
 
@@ -402,60 +145,5 @@ func TestGetPendingFilesDrains(t *testing.T) {
 	}
 	if got := a.GetPendingFiles(); got != nil {
 		t.Fatalf("second pull = %v, want nil", got)
-	}
-}
-
-// TestRequireRelay pins the transfer-time half of issue #281. Hide my IP forces
-// the relay path, so without a TURN URL in the list the attempt gathers no
-// usable candidate and dies about thirty seconds later as a generic timeout,
-// which errors.ts turns into advice about both devices being online.
-//
-// The first case is the one a careless implementation breaks: a STUN-only
-// server is a perfectly good server when the switch is off, and must not be
-// refused.
-func TestRequireRelay(t *testing.T) {
-	cases := []struct {
-		name      string
-		hideIP    bool
-		hasRelay  bool
-		degraded  bool
-		wantError error
-	}{
-		{"a relay-less server is fine with the switch off", false, false, false, nil},
-		{"a relay-less server is refused with the switch on", true, false, false, errNoRelay},
-		{"a relay-capable server is fine with the switch on", true, true, false, nil},
-		{"a relay-capable server is fine with the switch off", false, true, false, nil},
-		// The list this side is holding is the STUN-only fallback, so "no relay"
-		// is this side's guess rather than the server's answer. Blaming the
-		// server's configuration would be a confident guess at a wrong cause,
-		// and the usual causes (a wrong address, an un-proxied /api/, the TURN
-		// endpoint's rate limiter) are all things the reader can act on.
-		{"a list that could not be read does not blame the server", true, false, true, errRelayUnknown},
-		// Degraded but a relay somehow present cannot happen through
-		// ice.FetchDetail today, since the fallback is STUN only. Pinned so a
-		// future fallback with a relay in it does not silently start refusing.
-		{"degraded is irrelevant once a relay is in the list", true, true, true, nil},
-	}
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			err := requireRelay(tc.hideIP, tc.hasRelay, tc.degraded)
-			if !errors.Is(err, tc.wantError) {
-				t.Fatalf("requireRelay(%v, %v, %v) = %v, want %v",
-					tc.hideIP, tc.hasRelay, tc.degraded, err, tc.wantError)
-			}
-			if err == nil {
-				return
-			}
-			// The message has to name the switch, or it is the same
-			// unactionable failure in different words.
-			if !strings.Contains(err.Error(), "Hide my IP") {
-				t.Errorf("error %q does not name the setting to turn off", err)
-			}
-			// errors.ts matches on this clause to pass the sentence through
-			// verbatim; see its PASSTHROUGH list.
-			if !strings.Contains(err.Error(), "needs a TURN relay") {
-				t.Errorf("error %q lost the clause errors.ts anchors on", err)
-			}
-		})
 	}
 }
